refactor(structure): move CommentsContext to singletons
