+++ packages/sanity/src/_singletons/index.ts
@@ -27,6 +27,7 @@ export * from './presence/FormFieldPresenceContext'
 export * from './router/RouterContext'
 export * from './store/_legacy/ResourceCacheContext'
 export * from './structure/comments/authoring-path/CommentsAuthoringPathContext'
+export * from './structure/comments/CommentsContext'
 export * from './structure/comments/components/pte/comment-input/CommentInputContext'
 export * from './structure/components/pane/PaneContext'
 export * from './structure/panes/document/DocumentPaneContext'
