+++ packages/sanity/src/_singletons/structure/comments/CommentsContext.ts
@@ -1,5 +1,8 @@
 import {createContext} from 'react'
 
-import {type CommentsContextValue} from './types'
+import {CommentsContextValue} from '../../../core/comments/context/comments/types'
 
+/**
+ * @internal
+ */
 export const CommentsContext = createContext<CommentsContextValue | null>(null)
