+++ packages/sanity/src/core/comments/context/comments/index.ts
@@ -1,2 +1,1 @@
-export * from './CommentsContext'
 export * from './CommentsProvider'
